# Optimize embedding layers