This PR fixes:
- Initial WTE embedding step now parallelized across non-semantic codebooks instead of serial. Results are numerically equivalent. This shaves off about 300µs per token: i.e. most of this section:
<img width="1214" alt="image" src="https://github.com/user-attachments/assets/b48ee407-5bf6-4a74-bb24-5432ab8ed297">

Is now somewhere around here (ignore exact numbers):
<img width="188" alt="image" src="https://github.com/user-attachments/assets/9148ed59-9d10-473e-8b18-2dc38ded64b2">

- Top-k option added for fast codebooks to increase performance. Set to a rather generous 256 by default to avoid impacting generation quality. Shaves off another few hundred microseconds. To turn this off, set `--top-k=1024`.
- Saving `out.npy` for first time now doesn't error
- EOS token now at temp=1 after I noticed clipping was too aggressive. Raise an issue if anyone wants this to be configurable

Totally unscientific performance speedup on RTX 4090 release binary:
BEFORE:
```
Loaded prompt with shape [9, 166]
166 prompt processing timesteps (1630.66 tokens/s)
\ Tokens: 127 [00:00:00] 131.9143/s iterations/s                                                                                                                                                                                          128 tokens generated (133.92 tokens/s, RTF: 3.109)
```

AFTER:
```
166 prompt processing timesteps (1656.16 tokens/s)
\ Tokens: 129 [00:00:00] 142.265/s iterations/s                                                                                                                                                                                           131 tokens generated (144.58 tokens/s, 6.917ms / token, RTF: 3.357)
```

Next steps: Kernel fusion on transformer layers